requirements.txt
click==7.1.2
docx-mailmerge==0.4.0
elementpath==1.3.3
et-xmlfile==1.0.1
Flask==1.0.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
lxml==4.6.3
jsonpickle==1.2
MarkupSafe==1.1.1
openpyxl==2.5.12
Werkzeug==1.0.1
xmlschema==1.0.18
requests-toolbelt==0.9.1

backports.functools-lru-cache
lazy-object-proxy
wrapt
colorama
configparser
isort>=4.2.5
mccabe

googleapis_common_protos
google-cloud-ndb

six
setuptools
moves
